Add common method for GMA and WWFF spotting
GMA spotting can also be extended to other activities in future
